ファイルリストをFilePathListコンポーネントに分割
